Refactor : 초기화 안되었던 변수들 초기화

+++ main.cpp
@@ -95,7 +95,7 @@ int APIENTRY WinMain(HINSTANCE _hInstance, HINSTANCE _hPrevInstance,
 	// 메인게임 해제
 	g_mainGame.Release();
 
-	return message.wParam;
+	return (int)message.wParam;
 }
 
 LRESULT CALLBACK WndProc(HWND hWnd, UINT iMessage, WPARAM wParam, LPARAM lParam)
